Fixing recursive deps
