+++ package.json
@@ -19,7 +19,8 @@
     "test:int": "vitest --config vitest.integration.ts",
     "lint": "biome lint",
     "check-types": "tsc --noEmit",
-    "ddb:start": "docker compose up -d dynamodb"
+    "ddb:start": "docker compose up -d dynamodb",
+    "circular": "npx madge --circular --ts-config ./tsconfig.json --extensions ts,tsx src/"
   },
   "keywords": ["dynamodb", "aws", "typescript", "database"],
   "author": "Maunder",
