+++ src/builders/query-builder.ts
@@ -18,6 +18,7 @@ import {
 } from "../conditions";
 import { Paginator } from "./paginator";
 import type { GSINames, TableConfig } from "../types";
+import type { QueryBuilderInterface } from "./builder-types";
 
 /**
  * Configuration options for DynamoDB query operations.
@@ -104,20 +105,20 @@ type QueryExecutor<T extends Record<string, unknown>> = (
  * - Search habitats by type or region
  * - List incidents by date range
  * - Retrieve feeding schedules
- * 
+ *
  * The builder supports:
  * - Complex filtering conditions
  * - Sorting and pagination
  * - Global Secondary Indexes
  * - Attribute projection
- * 
+ *
  * @example
  * ```typescript
  * // Find active carnivores
  * const result = await new QueryBuilder(executor, eq('species', 'Tyrannosaurus'))
  *   .filter(op => op.eq('status', 'ACTIVE'))
  *   .execute();
- * 
+ *
  * // Search habitats by security level
  * const result = await new QueryBuilder(executor, eq('type', 'CARNIVORE'))
  *   .useIndex('security-level-index')
@@ -125,18 +126,20 @@ type QueryExecutor<T extends Record<string, unknown>> = (
  *   .select(['id', 'capacity', 'currentOccupants'])
  *   .sortDescending()
  *   .execute();
- * 
+ *
  * // List recent incidents
  * const result = await new QueryBuilder(executor, eq('type', 'INCIDENT'))
  *   .useIndex('date-index')
  *   .filter(op => op.gt('severityLevel', 5))
  *   .paginate(10);
  * ```
- * 
+ *
  * @typeParam T - The type of items being queried
  * @typeParam TConfig - The table configuration type for type-safe GSI selection
  */
-export class QueryBuilder<T extends Record<string, unknown>, TConfig extends TableConfig = TableConfig> {
+export class QueryBuilder<T extends Record<string, unknown>, TConfig extends TableConfig = TableConfig>
+  implements QueryBuilderInterface<T, TConfig>
+{
   private readonly keyCondition: Condition;
   private options: QueryOptions = {};
   private selectedFields: Set<string> = new Set();
@@ -223,25 +226,25 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - Search habitats by security level
    * - Find incidents by date
    * - List feeding schedules by time
-   * 
+   *
    * @example
    * ```typescript
    * // Find all dinosaurs of a specific species
    * builder
    *   .useIndex('species-status-index')
    *   .filter(op => op.eq('status', 'ACTIVE'));
-   * 
+   *
    * // Search high-security habitats
    * builder
    *   .useIndex('security-level-index')
-   *   .filter(op => 
+   *   .filter(op =>
    *     op.and([
    *       op.gt('securityLevel', 8),
    *       op.eq('status', 'OPERATIONAL')
    *     ])
    *   );
    * ```
-   * 
+   *
    * @param indexName - The name of the GSI to use (type-safe based on table configuration)
    * @returns The builder instance for method chaining
    */
@@ -257,12 +260,12 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - Monitor critical security systems
    * - Track immediate habitat changes
    * - Verify containment protocols
-   * 
+   *
    * Note:
    * - Consistent reads are not available on GSIs
    * - Consistent reads consume twice the throughput
    * - Default is eventually consistent reads
-   * 
+   *
    * @example
    * ```ts
    * // Check immediate dinosaur status
@@ -270,14 +273,14 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   .filter(op => op.eq('status', 'ACTIVE'))
    *   .consistentRead()
    *   .execute();
-   * 
+   *
    * // Monitor security breaches
    * const result = await new QueryBuilder(executor, eq('type', 'SECURITY_ALERT'))
    *   .useIndex('primary-index')
    *   .consistentRead(isEmergencyMode)
    *   .execute();
    * ```
-   * 
+   *
    * @param consistentRead - Whether to use consistent reads (defaults to true)
    * @returns The builder instance for method chaining
    */
@@ -321,20 +324,20 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - Find habitats with specific conditions
    * - Search for security incidents
    * - Monitor feeding patterns
-   * 
+   *
    * @example
    * ```typescript
    * // Find aggressive carnivores
-   * builder.filter(op => 
+   * builder.filter(op =>
    *   op.and([
    *     op.eq('diet', 'CARNIVORE'),
    *     op.gt('aggressionLevel', 7),
    *     op.eq('status', 'ACTIVE')
    *   ])
    * );
-   * 
+   *
    * // Search suitable breeding habitats
-   * builder.filter(op => 
+   * builder.filter(op =>
    *   op.and([
    *     op.between('temperature', 25, 30),
    *     op.lt('currentOccupants', 3),
@@ -342,7 +345,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   ])
    * );
    * ```
-   * 
+   *
    * @param condition - Either a Condition object or a callback function that builds the condition
    * @returns The builder instance for method chaining
    */
@@ -405,7 +408,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - Retrieve habitat statistics
    * - Monitor security metrics
    * - Optimize response size
-   * 
+   *
    * @example
    * ```typescript
    * // Get basic dinosaur info
@@ -415,7 +418,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   'stats.health',
    *   'stats.aggressionLevel'
    * ]);
-   * 
+   *
    * // Monitor habitat conditions
    * builder
    *   .select('securityStatus')
@@ -425,7 +428,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *     'lastInspectionDate'
    *   ]);
    * ```
-   * 
+   *
    * @param fields - A single field name or an array of field names to return
    * @returns The builder instance for method chaining
    */
@@ -474,7 +477,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - View incidents chronologically
    * - Track feeding schedule progression
    * - Monitor habitat inspections
-   * 
+   *
    * @example
    * ```typescript
    * // List dinosaurs by age
@@ -482,14 +485,14 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   .useIndex('age-index')
    *   .sortAscending()
    *   .execute();
-   * 
+   *
    * // View incidents chronologically
    * const result = await new QueryBuilder(executor, eq('type', 'SECURITY_BREACH'))
    *   .useIndex('date-index')
    *   .sortAscending()
    *   .execute();
    * ```
-   * 
+   *
    * @returns The builder instance for method chaining
    */
   sortAscending(): QueryBuilder<T> {
@@ -513,7 +516,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   .sortDescending()
    *   .limit(10)
    *   .execute();
-   * 
+   *
    * // Check latest dinosaur activities
    * const result = await new QueryBuilder(executor, eq('species', 'Velociraptor'))
    *   .useIndex('activity-time-index')
@@ -536,12 +539,12 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - View habitat inspection history
    * - Monitor security incidents
    * - Track feeding patterns
-   * 
+   *
    * The paginator handles:
    * - Tracking the last evaluated key
    * - Managing page boundaries
    * - Respecting overall query limits
-   * 
+   *
    * @example
    * ```typescript
    * // List dinosaurs by species
@@ -549,20 +552,20 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   .filter(op => op.eq('status', 'ACTIVE'))
    *   .useIndex('species-index')
    *   .paginate(10);
-   * 
+   *
    * // Process pages of security incidents
    * const paginator = new QueryBuilder(executor, eq('type', 'SECURITY_BREACH'))
    *   .filter(op => op.gt('severityLevel', 7))
    *   .sortDescending()
    *   .paginate(25);
-   * 
+   *
    * while (paginator.hasNextPage()) {
    *   const page = await paginator.getNextPage();
    *   console.log(`Processing incidents page ${page.page}, count: ${page.items.length}`);
    *   // Handle security incidents
    * }
    * ```
-   * 
+   *
    * @param pageSize - The number of items to return per page
    * @returns A Paginator instance that manages the pagination state
    * @see Paginator for more pagination control options
@@ -578,10 +581,10 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - Resume habitat inspection reviews
    * - Continue security incident analysis
    * - Store query position between sessions
-   * 
+   *
    * Note: This method is typically used for manual pagination.
    * For automatic pagination, use the paginate() method instead.
-   * 
+   *
    * @example
    * ```typescript
    * // First batch of dinosaurs
@@ -589,7 +592,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   .filter(op => op.eq('status', 'ACTIVE'))
    *   .limit(5)
    *   .execute();
-   * 
+   *
    * if (result1.lastEvaluatedKey) {
    *   // Continue listing dinosaurs
    *   const result2 = await new QueryBuilder(executor, eq('species', 'Velociraptor'))
@@ -597,11 +600,11 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *     .startFrom(result1.lastEvaluatedKey)
    *     .limit(5)
    *     .execute();
-   * 
+   *
    *   console.log('Additional dinosaurs:', result2.items);
    * }
    * ```
-   * 
+   *
    * @param lastEvaluatedKey - The exclusive start key from a previous query result
    * @returns The builder instance for method chaining
    */
@@ -617,35 +620,35 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - Check multiple habitat conditions
    * - Monitor various security levels
    * - Create report templates
-   * 
+   *
    * This is particularly useful when:
    * - Implementing pagination (used internally by paginate())
    * - Creating query templates
    * - Running multiple variations of a query
-   * 
+   *
    * @example
    * ```typescript
    * // Create base dinosaur query
    * const baseQuery = new QueryBuilder(executor, eq('species', 'Velociraptor'))
    *   .useIndex('status-index')
    *   .select(['id', 'status', 'location']);
-   * 
+   *
    * // Check active dinosaurs
    * const activeRaptors = baseQuery.clone()
    *   .filter(op => op.eq('status', 'HUNTING'))
    *   .execute();
-   * 
+   *
    * // Check contained dinosaurs
    * const containedRaptors = baseQuery.clone()
    *   .filter(op => op.eq('status', 'CONTAINED'))
    *   .execute();
-   * 
+   *
    * // Check sedated dinosaurs
    * const sedatedRaptors = baseQuery.clone()
    *   .filter(op => op.eq('status', 'SEDATED'))
    *   .execute();
    * ```
-   * 
+   *
    * @returns A new QueryBuilder instance with the same configuration
    */
   clone(): QueryBuilder<T, TConfig> {
@@ -662,17 +665,17 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    * - Check habitat conditions
    * - Monitor security incidents
    * - Track feeding patterns
-   * 
+   *
    * The method returns both the matched items and, if there are more results,
    * a lastEvaluatedKey that can be used with startFrom() to continue the query.
-   * 
+   *
    * @example
    * ```typescript
    * try {
    *   // Find active carnivores in specific habitat
    *   const result = await new QueryBuilder(executor, eq('habitatId', 'PADDOCK-A'))
    *     .useIndex('species-status-index')
-   *     .filter(op => 
+   *     .filter(op =>
    *       op.and([
    *         op.eq('diet', 'CARNIVORE'),
    *         op.eq('status', 'ACTIVE'),
@@ -682,9 +685,9 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *     .sortDescending()
    *     .limit(5)
    *     .execute();
-   * 
+   *
    *   console.log(`Found ${result.items.length} dangerous dinosaurs`);
-   * 
+   *
    *   if (result.lastEvaluatedKey) {
    *     console.log('Additional threats detected');
    *   }
@@ -692,7 +695,7 @@ export class QueryBuilder<T extends Record<string, unknown>, TConfig extends Tab
    *   console.error('Security scan failed:', error);
    * }
    * ```
-   * 
+   *
    * @returns A promise that resolves to an object containing:
    *          - items: Array of items matching the query
    *          - lastEvaluatedKey: Token for continuing the query, if more items exist
